hack: reuse same rbBitmap for DL retransmission
The libscheduler does not return a correct rbBitmap
(as far as I understand).

Let's memorize the one used for 1st transmission and reuse it.

If there are two or more UEs in the system then things may
fail if the scheduler allocates the PRBs used by UE 1 in 1st
transmission to UE 2 at the same time of a retransmission for
UE 1.